Update
chpa14 > sec03 > Quiz1, Quiz2(FileManager) update

+++ chap14/src/sec03/exam03/quiz/FileManager.java
@@ -5,16 +5,20 @@ import java.io.File;
 public class FileManager {
 	
 	public void deleteFile(File dir) {
-		File[] files = dir.listFiles();
-		for (File file : files) {
-			if (file.isDirectory()) {
-				System.out.println(file.getPath() +" -> 삭제 안됨");
-			} else {
-				if (file.delete()) {
-					System.out.println(file.getPath() +" -> 삭제 완료");					
+		
+		if (dir.exists() && dir.isDirectory()) {
+			File[] files = dir.listFiles();
+			for (File file : files) {
+				if (file.isDirectory()) {
+					System.out.println(file.getPath() +" -> 삭제 안됨");
+				} else {
+					if (file.delete()) {
+						System.out.println(file.getPath() +" -> 삭제 완료");					
+					}
 				}
 			}
 		}
+		else System.out.println("폴더가 아닙니다.");
 	}
 
 }
